Fold profile traits into class
They add complexity and there is little reason to believe they can be
used for other models.
